refactor: card page refactoring
