+++ src/pages/card-page/card-page.tsx
@@ -1,37 +1,16 @@
-import { useState } from 'react'
-import { Link, generatePath, useParams } from 'react-router-dom'
+import { Link } from 'react-router-dom'
 
 import { ArrowBackOutline } from '@/assets/icons'
 import { LearnCard } from '@/components/ui/layout-components'
 import { Button } from '@/components/ui/primitives'
 import { cn } from '@/pages/card-page/card-page.styles'
-import { useGetDeckQuery, useGetRandomCardQuery, useSaveGradeOfCardMutation } from '@/services'
-import { PATH } from '@/shared/enums'
+import { useCardPageData } from '@/shared/hooks'
 import { FlexContainer } from '@/shared/ui/flex-container'
 import { Page } from '@/shared/ui/page'
 
 export const CardPage = () => {
-  const { deckId } = useParams()
-  const deckPagePath = deckId && generatePath(PATH.DECK, { deckId })
-
-  const { data: deck, isFetching: isDeckFetching } = useGetDeckQuery({ id: deckId ?? '' })
-  const { data: randomCard, isLoading: isRandomCardFetching } = useGetRandomCardQuery({
-    id: deckId ?? '',
-  })
-
-  const [showAnswer, setShowAnswer] = useState(false)
-
-  const [saveGrade, { isLoading: isSaveGradeLoading }] = useSaveGradeOfCardMutation()
-
-  const nextQuestionHandler = async (cardId: string, grade: number) => {
-    const result = await saveGrade({ cardId, grade, id: deck?.id ?? '' })
-
-    if (result) {
-      setShowAnswer(false)
-    }
-  }
-
-  const isLoad = isDeckFetching || isRandomCardFetching || isSaveGradeLoading
+  const { deck, deckPagePath, isLoad, nextQuestionHandler, randomCard, setShowAnswer, showAnswer } =
+    useCardPageData()
 
   return (
     <Page load={isLoad}>
